Crate precompile_utils

Modules

costs

Cost calculations. TODO: PR EVM to make those cost calculations public.

data

handle

logs

modifier

Provide checks related to function modifiers (view/payable).

precompile_set

Provide utils assemble precompiles and precompilesets into a final precompile set with security checks. All security checks are enabled by default and must be disabled explicely throught type annotations.

prelude

substrate

Utils related to Substrate features:

testing

Traits

StatefulPrecompile

Trait similar to fp_evm::Precompile but with a &self parameter to manage some state (this state is only kept in a single transaction and is lost afterward).

Functions

error

Return an error with provided (static) text. Using the revert function of Gasometer is preferred as erroring consumed all the gas limit and the error message is not easily retrievable.

revert

succeed

Type Definitions

EvmResult

Alias for Result returning an EVM precompile error.