Module precompile_utils::data

Re-exports

pub use bytes::*;

pub use native::*;

Modules

bytes

native

xcm

Encoding of XCM types for solidity

Macros

paste

Structs

EvmDataReader

Wrapper around an EVM input slice, helping to parse it. Provide functions to parse common types.

EvmDataWriter

Help build an EVM input/output data.

SolidityConvert

Adapter to parse data as a first type then convert it to another one. Useful for old precompiles in which Solidity arguments where set larger than the needed Rust type.

String

A UTF-8–encoded, growable string.

Traits

EvmData

Data that can be converted from and to EVM data types.