Trait precompile_utils::precompile_set::PrecompileSetFragment

pub trait PrecompileSetFragment {
    fn new() -> Self;

    fn execute(
        &self, 
        handle: &mut impl PrecompileHandle
    ) -> Option<PrecompileResult>;

    fn is_precompile(&self, address: H160) -> bool;

    fn used_addresses(&self) -> Vec<H160>;
}

A fragment of a PrecompileSet. Should be implemented as is it was a PrecompileSet containing only the precompile(set) it wraps. They can be combined into a real PrecompileSet using PrecompileSetBuilder.

Required methods

fn new() -> Self

Instanciate the fragment.

fn execute(
&self,
handle: &mut impl PrecompileHandle
) -> Option<PrecompileResult>

Execute the fragment.

fn is_precompile(&self, address: H160) -> bool

Is the provided address a precompile in this fragment?

fn used_addresses(&self) -> Vec<H160>

Return the list of addresses covered by this fragment.