Trait precompile_utils::revert::BacktraceExt

pub trait BacktraceExt {
    fn map_in_tuple_to_field(self, fields: &[&'static str]) -> Self;
}

Additional function for everything having a Backtrace.

Required Methods

fn map_in_tuple_to_field(self, fields: &[&'static str]) -> Self

Map last tuple entry into a field. Does nothing if last entry is not a tuple. As in Solidity structs are equivalent to tuples and are tricky to parse correctly, it allows to parse any struct as a tuple (with the correct implementation in this crate) and then map tuple indices to struct fields.