Enum precompile_utils::solidity::revert::RevertReason

#[non_exhaustive]
pub enum RevertReason {
    Custom(String),
    ReadOutOfBounds {
        what: String,
    },
    UnknownSelector,
    ValueIsTooLarge {
        what: String,
    },
    PointerToOutofBound,
    CursorOverflow,
    ExpectedAtLeastNArguments(usize),
}

Possible revert reasons.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Custom(String)

A custom revert reason if other variants are not appropriate.

ReadOutOfBounds

Fields

what: String

What was being read?

Tried to read data out of bounds.

UnknownSelector

An unknown selector has been provided.

ValueIsTooLarge

Fields

what: String

What was being read?

A value is too large to fit in the wanted type. For security reasons integers are always parsed as uint256 then casted to the wanted type. If the value overflows this type then this revert is used.

PointerToOutofBound

A pointer (used for structs and arrays) points out of bounds.

CursorOverflow

The reading cursor overflowed. This should realistically never happen as it would require an input of length larger than 2^64, which would cost too much to be included in a block.

ExpectedAtLeastNArguments(usize)

Used by a check that the input contains at least N static arguments. Often use to return early if the input is too short.

Implementations

impl RevertReason

pub fn custom(s: impl Into<String>) -> Self

Create a RevertReason::Custom from anything that can be converted to a String. Argument is the custom revert message.

pub fn read_out_of_bounds(what: impl Into<String>) -> Self

Create a RevertReason::ReadOutOfBounds from anything that can be converted to a String. Argument names what was expected to be read.

pub fn value_is_too_large(what: impl Into<String>) -> Self

Create a RevertReason::ValueIsTooLarge from anything that can be converted to a String. Argument names what was expected to be read.