Trait precompile_utils::Precompile

pub trait Precompile {
    fn execute(
        handle: &mut impl PrecompileHandle
    ) -> Result<PrecompileOutput, PrecompileFailure>;
}

One single precompile used by EVM engine.

Required Methods

fn execute(
handle: &mut impl PrecompileHandle
) -> Result<PrecompileOutput, PrecompileFailure>

Try to execute the precompile with given handle which provides all call data and allow to register costs and logs.