Trait precompile_utils::StatefulPrecompile

pub trait StatefulPrecompile {
    fn new() -> Self;
    fn execute(
        &self,
        handle: &mut impl PrecompileHandle
    ) -> EvmResult<PrecompileOutput>;
}

Trait similar to fp_evm::Precompile but with a &self parameter to manage some state (this state is only kept in a single transaction and is lost afterward).

Required Methods

fn new() -> Self

Instanciate the precompile. Will be called once when building the PrecompileSet at the start of each Ethereum transaction.

fn execute(
&self,
handle: &mut impl PrecompileHandle
) -> EvmResult<PrecompileOutput>

Execute the precompile with a reference to its state.