Crate precompile_utils

Modules

precompile_set

Provide utils assemble precompiles and precompilesets into a final precompile set with security checks. All security checks are enabled by default and must be disabled explicely throught type annotations.

testing

Macros

keccak256

Structs

Address

The address type of Solidity. H160 could represent 2 types of data (bytes20 and address) that are not encoded the same way. To avoid issues writing H160 is thus not supported.

Bytes

The bytes/string type of Solidity. It is different from Vec<u8> which will be serialized with padding for each u8 element of the array, while Bytes is tightly packed.

EvmDataReader

Wrapper around an EVM input slice, helping to parse it. Provide functions to parse common types.

EvmDataWriter

Help build an EVM input/output data.

LogsBuilder

Builder for PrecompileOutput.

RuntimeHelper

Helper functions requiring a Runtime. This runtime must of course implement pallet_evm::Config.

Enums

FunctionModifier

Represents modifiers a Solidity function can be annotated with.

Traits

EvmData

Data that can be converted from and to EVM data types.

LogExt

Extension trait allowing to record logs into a PrecompileHandle.

Precompile

One single precompile used by EVM engine.

PrecompileHandleExt

StatefulPrecompile

Trait similar to fp_evm::Precompile but with a &self parameter to manage some state (this state is only kept in a single transaction and is lost afterward).

Functions

call_cost

error

Return an error with provided (static) text. Using the revert function of Gasometer is preferred as erroring consumed all the gas limit and the error message is not easily retrievable.

log_costs

revert

succeed

Type Definitions

EvmResult

Alias for Result returning an EVM precompile error.

Attribute Macros

generate_function_selector

This macro allows to associate to each variant of an enumeration a discriminant (of type u32 whose value corresponds to the first 4 bytes of the Hash Keccak256 of the character string indicated by the user of this macro.