Crate precompile_utils

Modules

evm

precompile_set

Provide utils to assemble precompiles and precompilesets into a final precompile set with security checks. All security checks are enabled by default and must be disabled explicely throught type annotations.

prelude

solidity

Provides utilities for compatibility with Solidity tooling.

substrate

Utils related to Substrate features:

testing

Macros

assert_event_emitted

Panics if an event is not found in the system log of events

assert_event_not_emitted

keccak256

mock_account

Traits

Precompile

One single precompile used by EVM engine.

Type Aliases

EvmResult

Alias for Result returning an EVM precompile error.

Attribute Macros

precompile

precompile_name_from_address