Module precompile_utils::prelude

Re-exports

pub use crate::data::Address;

pub use crate::data::BoundedBytes;

pub use crate::data::BoundedString;

pub use crate::data::BoundedVec;

pub use crate::data::EvmData;

pub use crate::data::EvmDataReader;

pub use crate::data::EvmDataWriter;

pub use crate::data::SolidityConvert;

pub use crate::data::UnboundedBytes;

pub use crate::data::UnboundedString;

pub use crate::handle::PrecompileHandleExt;

pub use crate::logs::log0;

pub use crate::logs::log1;

pub use crate::logs::log2;

pub use crate::logs::log3;

pub use crate::logs::log4;

pub use crate::logs::LogExt;

pub use crate::modifier::check_function_modifier;

pub use crate::modifier::FunctionModifier;

pub use crate::read_args;

pub use crate::read_struct;

pub use crate::revert;

pub use crate::revert;

pub use crate::revert::BacktraceExt;

pub use crate::revert::InjectBacktrace;

pub use crate::revert::MayRevert;

pub use crate::revert::Revert;

pub use crate::revert::RevertExt;

pub use crate::revert::RevertReason;

pub use crate::substrate::RuntimeHelper;

pub use crate::substrate::TryDispatchError;

pub use crate::succeed;

pub use crate::EvmResult;

pub use crate::StatefulPrecompile;

Macros

keccak256

Structs

PrecompileOutput

Data returned by a precompile on success.

Traits

PrecompileHandle

Handle provided to a precompile to interact with the EVM.

Attribute Macros

generate_function_selector

This macro allows to associate to each variant of an enumeration a discriminant (of type u32 whose value corresponds to the first 4 bytes of the Hash Keccak256 of the character string indicated by the user of this macro.

precompile